openstack_dashboard/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
openstacksdk==1.3.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
redis==4.6.0
celery==5.3.4
